chore: move LttngEventFilter to lttng_event_filter.py
